Implemented centralized URL map and 'api' directory
Created a separate 'flask_app/api/' directory for organizing Flask-related
files. Created centralized URL map at 'flask_app/api/views.py'. Organized
related scripts into dedicated files on the same directory.
