+++ flask_app/api/config.py
@@ -1,0 +1,2 @@
+ALLOWED_EXTENSIONS = {'csv'}
+UPLOAD_FOLDER = 'data/'
