+++ flask_app/api/utils.py
@@ -1,0 +1,5 @@
+from flask_app.api.config import ALLOWED_EXTENSIONS
+
+
+def allowed_file(filename):
+    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
